refactor(log-guard): isolate the canonical logs schema check (split S12 L1/3)
